requirements.txt
# Core dependencies
numpy==1.24.3
opencv-python==4.8.0.76
pytesseract==0.3.10
PyYAML==6.0.1
pdf2image==1.16.3
Pillow==10.0.0

# Test dependencies
pytest==7.4.0
pytest-cov==4.1.0

# Additional dependencies
python-dotenv 